Smooth scrolling
Instead of clearing the entire screen every time we need to refresh the
screen, we're now jumping to the top of the screen and clearing only the
line we're working with. This drastically cuts down on flickering
because we only ever have one blank line instead of a blank screen that
is filled back in.